Recover AEM and AEI resource pipelines

+++ src/engine/render/Engine.cpp
@@ -1537,7 +1537,7 @@ unsigned char AbyssEngine::Engine::enableShader;
 unsigned char AbyssEngine::Engine::EnableRefract;
 void *AbyssEngine::Engine::LodDistShader;
 unsigned char AbyssEngine::Engine::DisableRefract;
-int AbyssEngine::Engine::AnisotropyValue;
+float AbyssEngine::Engine::AnisotropyValue;
 unsigned char AbyssEngine::Engine::KeepRawMeshData;
 unsigned char AbyssEngine::Engine::backupSaveGames;
 int AbyssEngine::Engine::MultiSampleValue;
